requirements.txt
python-socketio[asyncio_client]==5.1.0
requests
